backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pandas
openpyxl      # pour lire les fichiers Excel .xls